requirements.txt
Package     Version
----------- -------
Pillow      7.0.0  
pip         20.0.2 
pytesseract 0.3.2  
setuptools  45.2.0 
wheel       0.34.2 
